+++ resources/js/main.js
@@ -1,7 +1,0 @@
-import AppStyles from '../scss/main.scss'
-
-import * as GarnishExportModal from './components/GarnishExportModal';
-import ExportButtonComponent from './components/ExportButtonComponent';
-
-let exportButtonComponent = new ExportButtonComponent();
-exportButtonComponent.init();