+++ src/web/assets/src/js/main.js
@@ -1,0 +1,6 @@
+import '../css/main.scss';
+import './components/GarnishExportModal';
+import ExportButtonComponent from './components/ExportButtonComponent';
+
+const exportButtonComponent = new ExportButtonComponent();
+exportButtonComponent.init();